Refactor logic from Game into Loading
Game has a few state variables that cannot be maintained there. Having it there causes issues when it rerenders, since the re-rendering trickles down to Loading and causes the Foodish API to get inundated with calls.

For now, all we need it to do is:
- Delegate Loading Screen to get memory cards; this means moving the types of cards + types for current level out.
- Load memory cards onto the screen.

And also we split other parts of the state like GAME_STATE and currentState into the separate GameManager.