Update naming on memory card and add ID attribute
The ID is to cross-check the card with the list of stored card statuses in Game.js and see if it has already been selected.

The naming pulls out the end of the filepath; the prior logic split the filepath where any numbers were found. This splits by '/' and then does the number split to make sure we get the last piece of the filepath.