Fix timer expiring causing exception
The exception was because of a re-render of the timer component conflicting with that of the parent GameManager, so instead what we had to do was set the state of variable `timerExpired` to true. Then, only on after the next re-render, would it unload the prop rather than right in the middle of it.